albedo/SDDNet/requirements.txt
numpy>=1.19.0
tqdm>=4.62.0
opencv-python>=4.5.0
pillow>=8.0.0
pandas>=1.3.0
configargparse>=1.5.0
efficientnet-pytorch>=0.7.0
ptflops>=0.6.0
pydensecrf>=1.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
thop>=0.1.0